+++ src/main/java/com/wex/currencyconverter/dto/ConvertedPurchaseTransactionDTO.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -17,7 +18,7 @@ public class ConvertedPurchaseTransactionDTO {
     private long id;
     private String description;
     private LocalDate transactionDate;
-    private Double usDollarAmount;
+    private BigDecimal usDollarAmount;
 
     private List<ConvertedAmountDTO> convertedAmounts;
