+++ src/main/java/com/wex/currencyconverter/model/PurchaseTransaction.java
@@ -10,6 +10,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -24,7 +25,7 @@ public class PurchaseTransaction {
     private long id;
     private String description;
     private LocalDate transactionDate;
-    private Double purchaseAmount;
+    private BigDecimal purchaseAmount;
 
 
 }
